cityos_support/requirements.txt
elasticsearch
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
starlette
aiofiles
mysqlclient==2.1.1
requests
datetime
chardet
openpyxl
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-keycloak
pydantic-settings
paho-mqtt
python-dotenv
